Create opt-in for models
The project already has an --exclude option. Besides the
--exclude-option develop a --include-models supported: only looks at
certain dirs and models. Update the documentation and tests for this
feature.

When using --exclude, for each exception you need to remember to add it
to exclude, otherwise you will lose media files. The opposite effect is
very desirable.

Resolved create opt-in for models