Place files in quarantine
Place files first in the media_folder + "quarantine/" before removing.

This allows for a safer removal of the files. If the files are still
needed it can safely be copied back from quarantine to the media folder.
The next step is to run a cronjob to remove files that have been in
quarantine for 3 months.

Resolves moves files in quarantine