Update tests
Update the tests and move the move_to_quarantine method to the cleanup
class.

The previous commit did not include any test to the new feature. Also
some issues were raised like files with the same name moved to the
quarantine would overwrite the previous files. These issues are added
as test and resolved in the code.