Add command for removing files in quarantine
Add a command option to remove files that are 90 days or longer in
quarantine. Update the Readme and tests accordingly.

Since the previous 2 commits the files are first placed in quarantine
instead of being removed. The option was missing to cleanup the
quarantine folder. With this option the circle is complete.

Resolves add command for removing files in quarantine.